feat: Custom button
Button to recognise which field the player wants to play, currently only function that the click the button marks the field and outputs the chess figure if there is one.

Future:
The player is marked where he can move the piece and with a click on such a square the move is executed.